Bug fixes. Added help message on login.

+++ SpaceTraders Client/Models/ScanServerModels.cs
@@ -35,6 +35,6 @@ namespace SpaceTraders_Client.Models
         public string Name { get; set; }
         public int X { get; set; }
         public int Y { get; set; }
-        public int? AnsibleProgress { get; set; }
+        public double? AnsibleProgress { get; set; }
     }
 }
